requirements.txt
flask>=3.0.0
openpyxl>=3.1.0
pywebview>=6.0

